+++ src/test/kotlin/calculator/OperatorKoTest.kt
@@ -1,4 +1,4 @@
-package step2.calculator
+package calculator
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
